refactor page component
